Restart fpm during the promote stage
The symlink is created in the promote stage, so restart fpm right after
it to minimise 'Class not found' errors encountered during deploys.

Note that this is just a work-around. Ideally, we'd have the symlink
creation *inside* the restart routine. If this work-around helps, we can
think about doing it the right way.

Bug: T229015
Change-Id: Ifd27230786288a7fc3a6ef6a792000a37f05d376