Depool the node before the promote stage
Checks are always executed afer a stage has completed. In this concrete
instance, that means that depooling happens after Scap exchanges the old
and new code. This isn't relevant for Parsoid/JS, but in the case of
Parsoid/PHP that is too late: new code is already present before the
node is depooled. Hence, depool the node during the fetch stage, which
happens right before the promote stage.

Bug: T238748
Change-Id: I5cbecc26700c7db4f84fd83715b2e6c9653cc5a0